extract `lib.user.passwords`
