+++ pycroft/lib/user/_old.py
@@ -73,6 +73,7 @@ from .user_id import (
     decode_type2_user_id,
     check_user_id,
 )
+from .passwords import generate_wifi_password
 
 mail_confirm_url = os.getenv('MAIL_CONFIRM_URL')
 password_reset_url = os.getenv('PASSWORD_RESET_URL')
@@ -140,63 +141,11 @@ def get_user_sheet(sheet_id: int) -> bytes | None:
     return storage.data
 
 
-@with_transaction
-def reset_password(user: User, processor: User) -> str:
-    if not can_target(user, processor):
-        raise PermissionError("cannot reset password of a user with a"
-                              " greater or equal permission level.")
-
-    plain_password = user_helper.generate_password(12)
-    user.password = plain_password
-
-    message = deferred_gettext("Password was reset")
-    log_user_event(author=processor,
-                   user=user,
-                   message=message.to_json())
-
-    return plain_password
-
 def can_target(user: User, processor: User) -> bool:
     if user != processor:
         return user.permission_level < processor.permission_level
     else:
         return True
-
-
-@with_transaction
-def reset_wifi_password(user: User, processor: User) -> str:
-    plain_password = generate_wifi_password()
-    user.wifi_password = plain_password
-
-    message = deferred_gettext("WIFI-Password was reset")
-    log_user_event(author=processor,
-                   user=user,
-                   message=message.to_json())
-
-    return plain_password
-
-
-def maybe_setup_wifi(user: User, processor: User) -> str | None:
-    """If wifi is available, sets a wifi password."""
-    if user.room and user.room.building.wifi_available:
-        return reset_wifi_password(user, processor)
-    return None
-
-
-@with_transaction
-def change_password(user: User, password: str) -> None:
-    # TODO: verify password complexity
-    user.password = password
-
-    message = deferred_gettext("Password was changed")
-    log_user_event(author=user,
-                   user=user,
-                   message=message.to_json())
-
-
-def generate_wifi_password() -> str:
-    return user_helper.generate_password(12)
-
 
 def create_user(
     name: str, login: str, email: str, birthdate: date,
@@ -1026,22 +975,6 @@ def send_password_reset_mail(user: User) -> bool:
 
     return True
 
-
-@with_transaction
-def change_password_from_token(token: str | None, password: str) -> bool:
-    if token is None:
-        return False
-
-    user = User.q.filter_by(password_reset_token=token).one_or_none()
-
-    if user:
-        change_password(user, password)
-        user.password_reset_token = None
-        user.email_confirmed = True
-
-        return True
-    else:
-        return False
 
 
 def find_similar_users(name: str, room: Room, ratio: float) -> Iterable[User]:
